requirements.txt
gurobipy==11.0.1
matplotlib==3.5.2
numpy==1.22.4
numpy==1.24.3
pandas==2.2.2
Pyomo==6.7.1
python_dateutil==2.8.2
scipy==1.8.1
seaborn==0.13.2
